requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pytesseract==0.3.13
requests==2.32.5
urllib3==2.5.0
